+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/Tabs/AddCloseInteractiveTab.test.jsx
@@ -3,14 +3,17 @@ import classNames from 'classnames/bind';
 import IconAddPerson from 'terra-icon/lib/icon/IconAddPerson';
 import IconSearch from 'terra-icon/lib/icon/IconSearch';
 import Tabs from 'terra-tabs';
+import Radio from 'terra-form-radio';
+import Field from 'terra-form-field';
+import Input from 'terra-form-input';
+import Fieldset from 'terra-form-fieldset';
 import TabContentTemplate from './TabContentTemplate';
-import styles from './common/TabExample.module.scss';
+import styles from './TabsTemplate.module.scss';
 
 const cx = classNames.bind(styles);
 
 let i = -1;
-
-const AddCloseTab = () => {
+function AddCloseTabInteractive() {
   const tabsExample = [];
   const [tabs, setTabs] = useState(tabsExample);
   const [activeKey, setActiveKey] = useState();
@@ -18,8 +21,13 @@ const AddCloseTab = () => {
   tabsExample.push(
     {
       label: 'Patient Information',
-      key: 'AddcloseLabelTab',
+      key: 'AddCloseTabInteractiveLabelTab',
       content: 'Patient Information',
+      children:
+  <>
+    <Radio id="default-radio" labelText="Default Radio" name="default" disabled />
+    <Radio id="default1-radio" labelText="Default Radio" name="default" />
+  </>,
     },
   );
 
@@ -28,8 +36,31 @@ const AddCloseTab = () => {
       label: 'Add Patient',
       icon: <IconAddPerson />,
       isIconOnly: true,
-      key: 'AddcloseIconTab',
+      key: 'AddCloseTabInteractiveIconTab',
       content: 'Add Patient',
+      children:
+  <>
+    <div>
+      <Fieldset
+        type="checkbox"
+        legend="Give your full name here"
+        name="children_present"
+        value="children_present"
+        error="All fields must be filled out"
+        required
+      >
+        <Field label="First" isInline required htmlFor="first">
+          <Input id="first" type="text" name="first" defaultValue="" />
+        </Field>
+        <Field label="Middle" isInline required htmlFor="middle">
+          <Input id="middle" type="text" name="middle" defaultValue="" />
+        </Field>
+        <Field label="Last" isInline required htmlFor="last">
+          <Input id="last" type="text" name="last" defaultValue="" />
+        </Field>
+      </Fieldset>
+    </div>
+  </>,
     },
   );
 
@@ -37,7 +68,7 @@ const AddCloseTab = () => {
     {
       label: 'TPA & Insurance',
       customDisplay: <div className={cx('custom-display')}>TPA & Insurance</div>,
-      key: 'AddcloseCustomTab',
+      key: 'AddCloseTabInteractiveCustomTab',
       content: 'TPA & Insurance',
     },
   );
@@ -45,7 +76,7 @@ const AddCloseTab = () => {
   tabsExample.push(
     {
       label: 'Disabled Tab',
-      key: 'AddcloseDisabledTab',
+      key: 'AddCloseTabInteractiveDisabledTab',
       icon: <IconSearch />,
       isDisabled: true,
     },
@@ -54,7 +85,7 @@ const AddCloseTab = () => {
   tabsExample.push(
     {
       label: 'List of Out-Patient Details Department Wise',
-      key: 'AddcloselongLabel',
+      key: 'AddCloseTabInteractivelongLabel',
       content: 'List of Out-Patient Details Department Wise',
     },
   );
@@ -62,7 +93,7 @@ const AddCloseTab = () => {
   tabsExample.push(
     {
       label: 'Ambulance Service',
-      key: 'AddcloselastTab',
+      key: 'AddCloseTabInteractivelastTab',
       content: 'Ambulance Service',
     },
   );
@@ -83,26 +114,28 @@ const AddCloseTab = () => {
       {
         label: `Added New Tab_${i}`,
         isIconOnly: false,
-        key: `AddcloseTab_${i}`,
+        key: `AddCloseTabInteractiveTab_${i}`,
         content: `Added New Tab_${i}`,
         isActive: true,
       },
     );
-    setActiveKey(`AddcloseTab_${i}`);
+    setActiveKey(`AddCloseTabInteractiveTab_${i}`);
     setTabs(tabsArray);
   };
 
   return (
-    <div className={cx('content-wrapper')}>
-      <Tabs activeKey={activeKey} setFocusOnContent isClosable onSelectAddButton={addMoreTabPanes} ariaLabelAddTab="Add Tab" onTabClose={handleTabClose}>
+    <div className={cx('content-wrapper-default')} id="tabs-container">
+      <Tabs activeKey={activeKey} isClosable onSelectAddButton={addMoreTabPanes} ariaLabelAddTab="Add Tab" onTabClose={handleTabClose}>
         { tabs.map((tab) => (
           <Tabs.Pane label={tab.label} isIconOnly={tab.isIconOnly} customDisplay={tab.customDisplay} isDisabled={tab.isDisabled} icon={tab.icon} key={tab.key} id={tab.key}>
-            <TabContentTemplate label={tab.content} />
+            <TabContentTemplate label={tab.content}>
+              {tab.children}
+            </TabContentTemplate>
           </Tabs.Pane>
         ))}
       </Tabs>
     </div>
   );
-};
+}
 
-export default AddCloseTab;
+export default AddCloseTabInteractive;