+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/Tabs/AddCloseTab.test.jsx
@@ -93,7 +93,7 @@ function AddTabandCloseTab() {
 
   return (
     <div className={cx('content-wrapper-default')} id="tabs-container">
-      <Tabs activeKey={activeKey} isClosable onSelectAddButton={addMoreTabPanes} ariaLabelAddTab="Add Tab" onTabClose={handleTabClose}>
+      <Tabs setFocusOnContent activeKey={activeKey} isClosable onSelectAddButton={addMoreTabPanes} ariaLabelAddTab="Add Tab" onTabClose={handleTabClose}>
         { tabs.map((tab) => (
           <Tabs.Pane label={tab.label} isIconOnly={tab.isIconOnly} customDisplay={tab.customDisplay} isDisabled={tab.isDisabled} icon={tab.icon} key={tab.key} id={tab.key}>
             <TabContentTemplate label={tab.content} />