+++ packages/terra-tabs/src/Tabs.jsx
@@ -79,6 +79,10 @@ const propTypes = {
    * The label to set on the add icon element.
    */
   ariaLabelAddTab: PropTypes.string,
+  /**
+   * Sets focus on content when set to `true`.
+   */
+  setFocusOnContent: PropTypes.bool,
 };
 
 const defaultProps = {
@@ -86,6 +90,7 @@ const defaultProps = {
   fill: false,
   isDraggable: false,
   isClosable: false,
+  setFocusOnContent: false,
 };
 
 class Tabs extends React.Component {
@@ -127,6 +132,7 @@ class Tabs extends React.Component {
       isDraggable,
       onTabOrderChange,
       isClosable,
+      setFocusOnContent,
       ...customProps
     } = this.props;
 
@@ -140,7 +146,7 @@ class Tabs extends React.Component {
         content = React.Children.map(child.props.children, contentItem => (
           React.cloneElement(contentItem)
         ));
-        tabContent = <CommonTabContent variant="framework">{content}</CommonTabContent>;
+        tabContent = <CommonTabContent setFocusOnContent={setFocusOnContent} variant="framework">{content}</CommonTabContent>;
       }
       commonTabItems.push(
         <CommonTabItem